+++ database/migrations/2023_03_21_061406_create_purchases_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('purchases', function (Blueprint $table) {
             $table->id();
             $table->integer('qty');
-            $table->float('price');
+            $table->decimal('price', 12, 2);
             $table->foreignId('product_id')->constrained();
             $table->timestamps();
         });
